ask_hr_agent/rag_service/requirements.txt
fastapi==0.123.10
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
google-adk==1.21.0
